Move everything into one file.
